# apps/api/openapi.yaml
openapi: "3.1.0"
info:
  title: ATHENA M&E Platform API
  version: "0.1.0"
  description: |
    Backend API for the ATHENA Monitoring & Evaluation platform.
    All endpoints (except /health and /auth/*) require a valid
    JWT issued via OAuth SSO, carried as an HttpOnly cookie
    (`athena_session`) or an `Authorization: Bearer <token>` header.
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://your.domain.com
    description: Production

# ── Security scheme ───────────────────────────────────────────
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: athena_session

  # ── Reusable schemas ────────────────────────────────────────
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UUID:
      type: string
      format: uuid
    ISOTimestamp:
      type: string
      format: date-time
    Role:
      type: string
      enum: [admin, supervisor, enumerator]
...
    Organization:
      type: object
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        name: { type: string }
        slug: { type: string }
        country: { type: string }
        created_at: { $ref: "#/components/schemas/ISOTimestamp" }
    User:
      type: object
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        org_id: { $ref: "#/components/schemas/UUID" }
        email: { type: string, format: email }
        display_name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        is_active: { type: boolean }
        created_at: { $ref: "#/components/schemas/ISOTimestamp" }
...
    QuarantineEntry:
      type: object
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        org_id: { $ref: "#/components/schemas/UUID" }
        raw_payload: { type: object, additionalProperties: true }
        entity_id: { $ref: "#/components/schemas/UUID", nullable: true }
        form_id: { $ref: "#/components/schemas/UUID", nullable: true }
        failure_reason:
          type: string
          enum:
            [freshness_violation, duplicate_entity, invalid_gps, schema_error]
        failure_detail:
          { type: object, additionalProperties: true, nullable: true }
        resolved: { type: boolean }
        queued_at: { $ref: "#/components/schemas/ISOTimestamp" }
...
    Pagination:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        pages: { type: integer }

  # ── Reusable responses ──────────────────────────────────────
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Insufficient role
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }

# ── Default security (override per-endpoint where needed) ─────
security:
  - sessionCookie: []

# ── Paths ─────────────────────────────────────────────────────
paths:
...
  /org:
    get:
      operationId: getOrg
      summary: Get organisation details
      tags: [Organisation]
      x-roles: [admin, supervisor]
      responses:
        "200":
          description: Organisation record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "401": { $ref: "#/components/responses/Unauthorized" }

    put:
      operationId: updateOrg
      summary: Update organisation name or slug
      tags: [Organisation]
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                slug: { type: string }
                country: { type: string }
      responses:
        "200":
          description: Updated organisation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /org/users:
    get:
      operationId: listUsers
      summary: List all users in the organisation
      tags: [Organisation]
      x-roles: [admin]
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

    post:
      operationId: inviteUser
      summary: Invite a new user (creates inactive record, sends email)
      tags: [Organisation]
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email: { type: string, format: email }
                role: { $ref: "#/components/schemas/Role" }
      responses:
        "201":
          description: User invited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /org/users/{id}/role:
    patch:
      operationId: updateUserRole
      summary: Change a user's role
      tags: [Organisation]
      x-roles: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/UUID" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { $ref: "#/components/schemas/Role" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /org/users/{id}:
    delete:
      operationId: deactivateUser
      summary: Deactivate a user (sets is_active = false)
      tags: [Organisation]
      x-roles: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/UUID" }
      responses:
        "204": { description: User deactivated }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ── Entities ─────────────────────────────────────────────────
...
  /quarantine:
    get:
      operationId: listQuarantine
      summary: List unresolved quarantine entries
      tags: [Quality]
      x-roles: [admin, supervisor]
      parameters:
        - name: failure_reason
          in: query
          schema:
            type: string
            enum:
              [freshness_violation, duplicate_entity, invalid_gps, schema_error]
        - name: form_id
          in: query
          schema: { $ref: "#/components/schemas/UUID" }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: Quarantine entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    {
                      type: array,
                      items: { $ref: "#/components/schemas/QuarantineEntry" },
                    }
                  pagination: { $ref: "#/components/schemas/Pagination" }
  /quarantine/{id}/resolve:
    post:
      operationId: resolveQuarantine
      summary: Approve a quarantine entry — re-runs DQA and writes to sector table
      tags: [Quality]
      x-roles: [admin, supervisor]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/UUID" }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                resolution_note: { type: string }
      responses:
        "200": { description: Entry resolved and written to sector table }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "404": { $ref: "#/components/responses/NotFound" }
  /quarantine/{id}/reject:
    post:
      operationId: rejectQuarantine
      summary: Permanently reject a quarantine entry
      tags: [Quality]
      x-roles: [admin, supervisor]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/UUID" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution_note]
              properties:
                resolution_note: { type: string }
      responses:
        "204": { description: Entry rejected }
        "404": { $ref: "#/components/responses/NotFound" }

  # ── Conflicts ─────────────────────────────────────────────────
...